Discard packets from different network
We need this code to make sure the right client received the packet. We
won't need to do this if there was any way to tell Mr. Unix that we are
only interested in receiving multicast packets on this socket from a
particular interface but AFAIK that is not possible, at least not in a
portable way.