Work-around some issues with gio on windows
glib uses a rather strange mechanism to init the winsock library (through
usage of g_inet_addres_*), so we better do it ourselves.

glib - rightly - doesn't set SO_REUSEADDR on g_socket_bind on windows
since it has different semantics. Nevertheless we need it on windows for
multicast sockets, so we add a helper function to set this socket option.