Recover from lost bye bye packets.
When ssdp server is forcefully stopped and started again,
it sends BYE BYE packets at its start.
But if server does not follow SSDP specification
or if we miss the BYE BYE packets for any other reason,
gssdp_resource_browser believes that the server was never restarted.
If location changed after server restart,
we are not able to talk to the server any more.

This patch checks whether location matches previous location.
If not, BYE BYE packet is simulated.

Compared to the previous code, ssdp:alive packets with no location
and no AL are dropped.

Signed-off-by: Branislav Katreniak <bkatreniak@nuvotechnologies.com>

https://bugzilla.gnome.org/show_bug.cgi?id=724030