gssdp-client: Fix potential strncpy() buffer overflow
strncpy() does not append a nul terminator if the input string is at
least as long as the output buffer length — so you must always reserve
space in the output buffer for a nul terminator. The req structure was
previously memset() to zero, so a nul terminator is present in arp_dev
already.

Spotted by Coverity (CID: 143731).

https://bugzilla.gnome.org/show_bug.cgi?id=764677